+++ bycycle/tests/features/test_shapes.py
@@ -1,12 +1,12 @@
-"""Tests the cycle-by-cycle shape feature computation function."""
+"""Tests for features.shape."""
 
 import pytest
 
 import numpy as np
 
 from bycycle.cyclepoints import find_extrema, find_zerox
-from bycycle.features.shape import (compute_shape_features, compute_cyclepoints, compute_durations,
-                                    compute_extrema_voltage, compute_symmetry, compute_band_amp)
+
+from bycycle.features.shape import *
 
 ###################################################################################################
 ###################################################################################################
@@ -74,24 +74,6 @@ def test_compute_shape_features(sim_args, find_extrema_kwargs, center_extrema, r
             np.testing.assert_allclose(df_shapes.loc[:, col], df_opp.loc[:, cols_trough[idx]])
         else:
             np.testing.assert_allclose(df_opp.loc[:, col], df_shapes.loc[:, cols_trough[idx]])
-
-
-def test_compute_cyclepoints(sim_args):
-
-    sig = sim_args['sig']
-    fs = sim_args['fs']
-    f_range = sim_args['f_range']
-
-    peaks, troughs = find_extrema(sig, fs, f_range)
-    rises, decays = find_zerox(sig, peaks, troughs)
-
-    df_samples = compute_cyclepoints(sig, fs, f_range)
-
-    assert (df_samples['sample_peak'] == peaks[1:]).all()
-    assert (df_samples['sample_zerox_decay'] == decays[1:]).all()
-    assert (df_samples['sample_zerox_rise'] == rises).all()
-    assert (df_samples['sample_last_trough'] == troughs[:-1]).all()
-    assert (df_samples['sample_next_trough'] == troughs[1:]).all()
 
 
 def test_compute_durations(sim_args):
