+++ bycycle/tests/plts/test_cyclepoints.py
@@ -1,9 +1,11 @@
-"""Tests plotting cyclepoints."""
+"""Tests for plts.cyclepoints."""
 
 from bycycle.cyclepoints import find_extrema, find_zerox
-from bycycle.plts.cyclepoints import plot_cyclepoints_df, plot_cyclepoints_array
-from bycycle.tests.utils import plot_test
+
+from bycycle.tests.tutils import plot_test
 from bycycle.tests.settings import TEST_PLOTS_PATH
+
+from bycycle.plts.cyclepoints import *
 
 ###################################################################################################
 ###################################################################################################
