+++ bycycle/tests/plts/test_features.py
@@ -1,11 +1,13 @@
-"""Tests plotting cycle features."""
+"""Tests plts.features."""
 
 import numpy as np
 
 from bycycle.burst import detect_bursts_cycles
-from bycycle.plts.features import plot_feature_hist, plot_feature_categorical
-from bycycle.tests.utils import plot_test
+
+from bycycle.tests.tutils import plot_test
 from bycycle.tests.settings import TEST_PLOTS_PATH
+
+from bycycle.plts.features import *
 
 ###################################################################################################
 ###################################################################################################
@@ -16,7 +18,6 @@ def test_plot_feature_hist(sim_args):
     df_features = sim_args['df_features']
     threshold_kwargs = sim_args['threshold_kwargs']
 
-    # Apply consistency burst detection
     df_features = detect_bursts_cycles(df_features, **threshold_kwargs)
 
     plot_feature_hist(df_features, 'amp_consistency', xlim=(0, 1), save_fig=True,
@@ -28,7 +29,6 @@ def test_plot_feature_categorical(sim_args):
     df_features = sim_args['df_features']
     threshold_kwargs = sim_args['threshold_kwargs']
 
-    # Apply consistency burst detection
     df_features = detect_bursts_cycles(df_features, **threshold_kwargs)
 
     # Compare first & second halves of the signal - distributions should be the same
